New user parameter to set the maximum number of steps in the solver
This parameter can be used to set the maximum number of steps to be
taken by the solver in its attempt to reach the next output time. If
set to a negative value, there is no maximum. It replaces the
CVODE_MXSTEPS parameter that was defined in astrochem.h. The default
value is unchanged, i.e. 1e6.